fix compilation

+++ src/defs.h
@@ -16,8 +16,8 @@ using AdjGraph = std::vector<std::vector<int>>;
 using NNVec = std::vector<std::pair<float, uint32_t>>;
 
 struct HNSWParameters {
-    int M = 16;
-    int ef_construction = 200;
+    size_t M = 16;
+    size_t ef_construction = 200;
 };
 
 using Duration = std::chrono::duration<double>;
